# 5.5x more CUDA performance with 5 minutes of work